requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
gradio==5.4.0
gradio_client==1.4.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
lightning==2.4.0
lightning-utilities==0.11.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
rsa==4.9
ruff==0.7.1
safehttpx==0.1.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
tomlkit==0.12.0
torch==2.5.1
torchmetrics==1.5.1
torchvision==0.20.1
tqdm==4.66.6
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
yarl==1.17.0